# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MaintainIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List technicians (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Technician list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianListResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create a technician account (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Technician created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get a technician's detail (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Technician detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update a technician (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Assets ----------
...
  /public/assets/{assetCode}:
    get:
      operationId: getPublicAsset
      tags: [public]
      summary: Get safe public asset info by asset code (no auth, for QR scans)
      parameters:
        - name: assetCode
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public asset info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAsset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /public/issues:
    post:
      operationId: createPublicIssue
      tags: [public]
      summary: Submit an issue report (no auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicIssueInput"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicIssueReceipt"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- AI ----------
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---------- Users / Auth ----------
...
    Technician:
      type: object
      required: [id, name, email, active, assignedIssuesCount, resolvedIssuesCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        active: { type: boolean }
        assignedIssuesCount: { type: integer }
        resolvedIssuesCount: { type: integer }
        createdAt: { type: string, format: date-time }
    TechnicianInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        phone: { type: string }
    TechnicianUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        active: { type: boolean }
    TechnicianListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Technician"

    # ---------- Assets ----------
...
    PublicHistoryEntry:
      type: object
      required: [date, action]
      properties:
        date: { type: string, format: date-time }
        action: { type: string }
        status: { type: ["string", "null"] }
    PublicAsset:
      type: object
      required:
        [id, name, assetCode, category, location, status, recentActivity]
      properties:
        id: { type: integer }
        name: { type: string }
        assetCode: { type: string }
        category: { type: string }
        location: { type: string }
        status:
          type: string
          enum: [operational, under_maintenance, out_of_service, retired]
        lastServiceDate: { type: ["string", "null"], format: date }
        nextServiceDate: { type: ["string", "null"], format: date }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/PublicHistoryEntry"

    # ---------- Issues ----------
...
    PublicIssueInput:
      type: object
      required:
        [
          assetCode,
          reporterName,
          reporterEmail,
          title,
          description,
          priority,
          category,
        ]
      properties:
        assetCode: { type: string }
        reporterName: { type: string, minLength: 1 }
        reporterEmail: { type: string }
        reporterPhone: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        priority: { type: string, enum: [low, medium, high, critical] }
        category: { type: string, minLength: 1 }
        imageUrls:
          type: array
          items: { type: string }
        aiTitle: { type: string }
        aiCategory: { type: string }
        aiPriority: { type: string, enum: [low, medium, high, critical] }
        aiPossibleCauses:
          type: array
          items: { type: string }
        aiDiagnosticChecks:
          type: array
          items: { type: string }
        aiSafetyNotes: { type: string }
        aiRecurringWarning: { type: string }
    PublicIssueReceipt:
      type: object
      required: [issueNumber, status]
      properties:
        issueNumber: { type: string }
        status: { type: string }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"